Add dynamic runtime configuration system
This commit sets up, and demonstrates, the use of runtime configurations
in the landing page site.

Configurations are stored in a YAML configuration file. The default
configuration for development is squareone.config.yaml.

This YAML file can be set, in production, as a ConfigMap-based file
that's loadable in the running container.

This data is loaded into the publicRuntimeConfig object in
next.config.js. This makes the data available to both server side and
client-side code.

When we implement secrets, the secrets can be loaded via the
serverRuntimeConfig as these are available *only* to server-side code.

Within a page component this data is