Create mock of /auth/api/v1/user-info endpoint
This is a mock of Gafaelfawr's user-info endpoint to demonstrate getting
login data.

The reason we need to mock it is because Gafaelfawr is challenging to
quickly deploy beside a development server (needs OAuth set up) and
needs to be on the same hostname as the user application in order to
share cookies.

The reason for the "rewrites" configuration is that next.js only allows
API endpoints to be in the api directory, but the Gafaelfawr route is
/auth/api/v1/user-info. The rewrite lets us internally make that URL
resolve to pages/api/user-info.js.

Finally, note that this mock API will never get called in production
because the ingress will ensure that /auth/ requests always go to
Gafaelfawr.