Added broadcastMarkdown to public config schema
This is a markdown-formatted message that can be set via the app's
configmap configuration that is displayed as a broadcast-style
notification (see https://sqr-060.lsst.io for info on RSP
notifications). Note that this will be replaced by the Semaphore API,
but it may be persistently useful as a fallback in case semaphore itself
is down.