Render configured broadcast markdown to HTML
This uses remark with the GitHub Flavored Markdown plugin to render the
broadcastMarkdown configuration field into HTML, and sets that HTML into
the BroadcastBanner component.

remark runs in node, so the conversion is done in the App component's
getInitialProps, which happens server-side since all pages use
getServerSideProps.

I tried to use jsdom and dompurify to sanitize the HTML, but could not
get those libraries to work in the context of getInitialProps. Since the
broadcast content comes from a trusted source via Kubernetes
configuration, I think this is fine for now.