Switch semaphore broadcasts to swr library
Now we can fully drop the custom fetch hook and use SWR. A bonus of this
approach is that SWR lets us easily fresh data on an interval (in this
case, every 60 s), allow us to have the homepage auto-update when
broadcasts change.

The useBroadcasts hook is special purpose to the BroadcastBannerStack
component cluster, but it could be moved out as a site-wide hook if the
data is needed elsewhere.