Add Times Square pages
This is ported from https://github.com/lsst-sqre/times-square-ui

The nb page is modified to not use the custom useRouter hook; let's see
if it's not needed after all.

In the future, we'll need to move the /nb/[nbSlug].js page to work with
the real URL layout for GitHub vs API-backed pages. We'll also need to
factor the big component out of /nb/[nbSlug].js and probably add it so
the TimesSquareViewer component or perhaps TimesSquarePage.