Refactor TSNotebookViewer into TimesSquarePage
The TimesSquarePage component provides a view into a single Times Square
page. It consists of a column showing page metadata/settings and a
column with the actual notebook content via the NotebookIframe
component (now a sub-component, and formerly the root component of
TimesSquareViewer).

This also removes a lot of the logic from the /nb/[nbSlug].js page,
which is useful because we'll have multiple page endpoints for different
Times Square UI contexts.