Refactor TimesSquareGitHubNav from TS app
The purpose of this refactor is to invert control over the sidebar so
that different Times Square app contexts, from different pages, can
insert different items in the sidebar. For example, the page navigation
component for the /github/ views should have this TimesSquareGitHubNav
component; user uploads will use a slightly different component.