Add a reset button
The reset handler provided by Formik resets the form to the initial
values. I'm not sure this is exactly what we want a global reset button
to do here through, since the initial values are that for the URL and
don't restore to parameter defaults.

Probably what we want is per-parameter reset affordances that undo
pending changes, and then make the global reset button resort the form
to the default view.