Set up storybook
This storybook set up is based on
https://storybook.js.org/blog/get-started-with-storybook-and-next-js/

There's a storybook issue with npm 8.6, and the recommended workaround
is to add legacy-peer-deps=true to .npmrc. This should be fixed
eventually.

Also, I needed to ignore some rules for eslint as anonymous default
exports are apparently best practice for storybook. Also trying ignore
the rule for react/no-unescaped-entities, though we'll see if that
becomes an issue later.