Move check run info to landing page from sidebar
On the landing page for check run, rather than include all the metadata
in the nav sidebar and push the listing of pages down, let's put that
info right on the landing page.

Add a showPrDetails prop to TimesSquarePrGitHubNav to control whether
the nav component includes that info in the sidebar.