Add next-mdx-remote
This enables to load any MDX source and render it in getStaticProps. For
example, this enables us to have certain pages be configured in YAML and
then rendered specially for that environment. The advantage of MDX in
this context is that we can include components like
headers/buttons/calls to action.

https://github.com/hashicorp/next-mdx-remote