Re-enable storybook for squareone app
Storybook was originally configured for the Squareone Next.js app, but
when we migrated to the monorepo, our hope was to have one Storybook for
all packages and apps; while this is possible, it's tricky to configure
right. Instead, we'll make a storybook for each project, and link them
together using Storybook's composition feature.