Enable light/dark theme in storybook
This puts a light/dark theme switcher in the storybook view. It
simulates the effect of changing data-theme attribute, much like
next-themes does.

I'm not sure this is the ultimate solution. Ideally we want to be able
to control the background of a component at the same time as we toggle
the theme. Or be able to show both light and dark themes at the same
time.