Create the global-css package
The intent of this package is to wrap up a reset, CSS custom properties
from Rubin Style Dictionary, and application of those tokens to HTML
elements into a single bundled styledsheet.

This replaces our use of styled-components global styles.

The first distributed output is for next.js applications but potentially
we could create specialized outputs for other projects.