Adopt global-css in Squared and Squareone
This drops the initiative to build a GlobalStyles component in squared
and instead switches us back to using a CSS import for the global style.

The main benefit of this final arrangement is that now we can but all
the business of integrating Rubin Style Dictionary components into the
global CSS into a dedicated CSS module, global-css, rather than needing
to do it in every application.

A limitation of this approach is that we're still not centralizing the
font import; but I'm not sure there's a good way to do that. We'll just
have to document how we load fonts. And with variable fonts this might
become easier too.