Compute login/out URLs with proper redirects
This adds the getLoginUrl and getLogoutUrl functions to compute those
auth URLs with redirects. The login URL will redirect back to the
current page while the logout URL redirects to the homepage, /.

Now the user menu component takes the current URL as a prop. This can be
obtained with the getCurrentUrl hook.

We're not integrating that hook into this story though because I don't
yet know how to handle mocking the react router for storybook.