Test opening the menu
This play function executes a click on the user button to expose the
menu. I was hoping to them also interrogate the user menu itself, but I
found out that radix is rendering the menu outside the scope of the
story! and therefore it isn't visible within the DOM in the canvas.