Implement ExecStats for page panel
This shows the execution status, including the age of the execution and
execution duration if available.

It's also where we put the button to request a recomputation of an
existing page.

The component's data is populated by the TimesSquareHtmlEventsContext.
We show how to mock the context in Storybook to design different states.