Create IconPill component in squared
This reusable component will provide a pill that contains an icon
alongside text content. It can be linked and the text and background
colours are configurable as props.

Make IconPill work with fontawesome icons

Uses can specify the icon as an array of group and icon set and name
strings, e.g. [ 'fas', 'book'].