Migrate squareone global css token to global-css
Initially we had globals.css in the Squareone app to provide CSS custom
properties / design tokens for the Next.js app. Now we want these tokens
to also be accessible to components built out from the Squared component
library, hence we're moving that CSS into global-css and distributing it
through the next.css file. This makes the token available to any Next.js
app in Squareone as well as in the storybook previews of Squared
components.

Eventually we'd like to move the management of many of these custom
properties up into rubin-style-dictionary but we'd have to figure out
how to manage things like elevation tokens.