Position the navigation menu under trigger
The Radix navigation menu currently has a flaw where the content
viewport is always centered under the navigation menu as a whole, rather
than under the trigger specifically. This isn't suitable for Squareone
because our navigation menu spans the entire window so the menu will
appear disassociated with the trigger.

This solution is from the GitHub issue thread and uses a
MutationObserver to update the position of the viewport.

https://github.com/radix-ui/primitives/issues/1462#issuecomment-2275683692

It has the side-effect right now of animating the menu sliding from the
centre to the trigger on first view. Perhaps we can tweak the animations
so this doesn't happen? Also, we might need to add code to
updatePosition that takes into account the browser viewport to ensure we
don't push menu content outside the view.