Adopt PrimaryNav in Header
First steps to adopting the PrimaryNav into the Header. There's a weird
bug where the Login component is being rendered server-side, and hence
the Logged-in state isn't always available.

This replaces the existing HeaderNav, which composed basic Next links in
a flex nav along with a GafaelfawrUserMenu with the PrimaryNav where the
User menu is now a regular menu item.