Handle viewport alignment near right side
Now we test whether the menu trigger is near the right side of the menu.
If it is, we switch to using the "right" for absolute alignment. If it
is not, we use the "left". In both cases, we use the absolute alignment
rather than doing transformX to avoid having the menu slide sideways,
which doesn't make sense since the menu now opens on click rather than
hover.