Add AppsMenu to HeaderNav
This is an initial implementation that demonstrates an Apps menu
that's displayed when the enableAppsMenu setting is added, and
shows the contents of the appLinks configuration.

This Apps menu implementation is challenging because it doesn't seem to
work well with the Next Link component, even if I use the asChild
composition method. I've tried to work around this by using a link
component that uses a span with a router.push onClick handler, but then
this breaks keyboard navigation. More work is needed.