Adopt the horizontal triad logo
This replaces the Rubin imagotype in the header with the horizontal
triad logo lockup. The big change here is that the logo is no longer in
line with the navigation components, and needs to be placed *above* the
main navigation in its own "PreLogo" container. Potentially we could put
other stuff on the right side of this PreLogo container;
rubinobservatory.org puts search up there.

The aria-label should help with the accessibility of the logo image
being part of the navigation.