Use CSS Modules in squared package
For Squared, we're exploring the use of CSS modules for resuable
components as a way of styling that doesn't have a runtime impact. We've
had small issues recently with styled-components SSR hydration and using
CSS modules will resolve that.

Squared will ship a styles.css export that consumer apps can import.
Shipping raw css allows us apps to use their own post processing and
treeshaking to optimize css that is shipped to clients.