Replace tsup in squared with app transpilation
Rather than having a separate build step for squared, this removes the
tsup build for squared and instead allows squared to be built directly by
the Next.js compiler using the transpilePackages configuration.

This shrewdly side-steps the issue of efficiently packaging squared.
We'd tried the "pizza slicing" method from
https://stackoverflow.com/a/73883783 but couldn't get the css modules to
correctly go along with the tsup build.

With transpilation, it also means that we get hot module reloading for
changes to squared.

If we ever have users of squared that are outside the squareone monorepo
then we will need to re-add a build process to squared (probably with
rollup, which has seems to have better css/css modules support).