Create DateTimePicker component
This component lets a user specify a datetime. The TextInput shows the
date time as an ISO8601 string. The user can click on a calendar icon to
get a popover where they can choose the date with a calendar component
(powered by react-day-picker) and additional controls for stepping the
hours and minutes. The user can also select the timezone for their
datetime. The popover is handled with radix-ui.