Add trailing action (button) for TextInput
This is similar to trailing icon (and in fact, they're mutually
exclusive props), the trailingAction adds an icon button to the right
side of the TextInput (within the visual focus ring of the input. An
application of this feature is to open a popover like for a date or time
picker.