Fix token created date showing "Expired"
The "Created" field in token details was incorrectly using
formatTokenExpiration(), which returns "Expired" for any date in the
past. Since creation dates are always in the past, they always
displayed as "Expired".

Added new formatTokenCreated() function that always returns the
ISO8601 timestamp format for creation dates, which is the expected
behavior for a timestamp that doesn't expire.

Updated TokenDetailsView component, Storybook stories, and tests to
use the new formatter.