Create Claude skills for domain knowledge
Claude skills is a new way of dynamically adding context to Claude when
needed, without having a bloated CLAUDE.md memory file. This introduces
several new Claude skills that are based on the original CLAUDE.md file
and also new ones that have come up as being useful, such as how to find
API docs for the science platform (platform-api-integration), and how to
use CSS variables in our design system (design-system skill). This has
let us significantly reduce the size of our CLAUDE file. As we work on
more features we can update and add to these skills to tweak Claude's
work.